Rename Accordion -> Disclosure, and AccordionGroup -> Accordion.
