+++ src/components/containers/Accordion/Accordion.stories.tsx
@@ -5,12 +5,9 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
 import * as React from 'react';
-
 import { LoremIpsum } from '../../../util/storybook/LoremIpsum.tsx';
-import { LayoutDecorator } from '../../../util/storybook/LayoutDecorator.tsx';
 
 import { Accordion } from './Accordion.tsx';
-import { OverflowTester } from '../../../util/storybook/OverflowTester.tsx';
 
 
 type AccordionArgs = React.ComponentProps<typeof Accordion>;
@@ -20,64 +17,70 @@ export default {
   component: Accordion,
   parameters: {
     layout: 'centered',
-    design: { type: 'figma', url: 'https://www.figma.com/design/ymWCnsGfIsC2zCz17Ur11Z/Design-System-UX?node-id=5634-155879&node-type=instance' },
   },
-  decorators: [
-    Story => <LayoutDecorator size="small"><Story/></LayoutDecorator>,
-  ],
   tags: ['autodocs'],
   argTypes: {
   },
+  decorators: [
+    Story => <div style={{ maxWidth: 500 }}><Story/></div>,
+  ],
   args: {
     children: (
-      <LoremIpsum paragraphs={1}/>
+      <>
+        <Accordion.Disclosure title="Title 1"><LoremIpsum/></Accordion.Disclosure>
+        <Accordion.Disclosure title="Title 2"><LoremIpsum/></Accordion.Disclosure>
+        <Accordion.Disclosure title="Title 3"><LoremIpsum/></Accordion.Disclosure>
+        <Accordion.Disclosure title="An accordion with a very long title that should cause text overflow">
+          <LoremIpsum paragraphs={3}/>
+        </Accordion.Disclosure>
+      </>
     ),
-    title: 'My Accordion',
   },
   render: (args) => <Accordion {...args}/>,
 } satisfies Meta<AccordionArgs>;
 
 
-export const Standard: Story = {};
-
-export const DefaultOpen: Story = {
+export const ExclusiveAccordion: Story = {
   args: {
-    open: true,
-    title: 'I should be open by default',
+    children: (
+      <>
+        <Accordion.Disclosure title="Only one of these accordions can be open">
+          <LoremIpsum/>
+        </Accordion.Disclosure>
+        <Accordion.Disclosure title="Title 2"><LoremIpsum/></Accordion.Disclosure>
+        <Accordion.Disclosure title="Title 3"><LoremIpsum/></Accordion.Disclosure>
+      </>
+    ),
+  }
+};
+export const NonexclusiveAccordion: Story = {
+  args: {
+    exclusive: false,
+    children: (
+      <>
+        <Accordion.Disclosure title="Multiple of these can be open at the same time">
+          <LoremIpsum short/>
+        </Accordion.Disclosure>
+        <Accordion.Disclosure open title="I am open by default"><LoremIpsum short/></Accordion.Disclosure>
+        <Accordion.Disclosure open title="I am also open by default"><LoremIpsum short/></Accordion.Disclosure>
+      </>
+    ),
   },
 };
 
-export const WithFocusClosed: Story = {
+export const WithFocus: Story = {
   args: {
-    open: false,
-    summaryProps: { className: 'pseudo-focus-visible' },
-    title: 'I should have a focus outline',
-    children: <>The border radius on the outline should be on both top and bottom when closed.</>,
-  },
-};
-
-export const WithFocusOpen: Story = {
-  args: {
-    open: true,
-    summaryProps: { className: 'pseudo-focus-visible' },
-    title: 'I should have a focus outline',
-    children: <>The border radius on the outline should be only on the top when open.</>,
-  },
-};
-
-export const WithTextOverflow: Story = {
-  args: {
-    title: 'An accordion with a long title that should cause text overflow',
-  },
-};
-
-// Note: interesting quirk with the open/close animation. Because it's the `::details-content` that gets its height
-// animated, if there is a lot of overflowing content, there is a visual delay in the close animation.
-export const WithScroll: Story = {
-  args: {
-    open: true,
-    title: 'An accordion with lots of content',
-    children: <LoremIpsum paragraphs={6}/>,
-    style: { maxHeight: '80svh' },
+    children: (
+      <>
+        <Accordion.Disclosure title="Title 1"><LoremIpsum/></Accordion.Disclosure>
+        <Accordion.Disclosure title="Title 2"><LoremIpsum/></Accordion.Disclosure>
+        <Accordion.Disclosure
+          title="I should have a focus outline"
+          summaryProps={{ className: 'pseudo-focus-visible' }}
+        >
+          <LoremIpsum/>
+        </Accordion.Disclosure>
+      </>
+    ),
   },
 };
